+++ Library/Homebrew/api_hashable.rb
