+++ Library/Homebrew/dev-cmd/bottle.rb
@@ -227,7 +227,7 @@ module Homebrew
       private
 
       sig {
-        params(string: String, keg: Keg, ignores: T::Array[String],
+        params(string: String, keg: Keg, ignores: T::Array[Regexp],
                formula_and_runtime_deps_names: T.nilable(T::Array[String])).returns(T::Boolean)
       }
       def keg_contain?(string, keg, ignores, formula_and_runtime_deps_names = nil)
@@ -373,35 +373,17 @@ module Homebrew
         [gnu_tar(gnu_tar_formula), reproducible_gnutar_args(mtime)].freeze
       end
 
-      sig { params(formula: T.untyped).returns(T::Array[T.untyped]) }
+      sig { params(formula: Formula).returns(T::Array[Regexp]) }
       def formula_ignores(formula)
-        ignores = []
-        cellar_regex = Regexp.escape(HOMEBREW_CELLAR)
-        prefix_regex = Regexp.escape(HOMEBREW_PREFIX)
-
         # Ignore matches to go keg, because all go binaries are statically linked.
         any_go_deps = formula.deps.any? do |dep|
           Version.formula_optionally_versioned_regex(:go).match?(dep.name)
         end
-        if any_go_deps
-          go_regex = Version.formula_optionally_versioned_regex(:go, full: false)
-          ignores << %r{#{cellar_regex}/#{go_regex}/[\d.]+/libexec}
-        end
+        return [] unless any_go_deps
 
-        # TODO: Refactor and move to extend/os
-        # rubocop:disable Homebrew/MoveToExtendOS
-        ignores << case formula.name
-        # On Linux, GCC installation can be moved so long as the whole directory tree is moved together:
-        # https://gcc-help.gcc.gnu.narkive.com/GnwuCA7l/moving-gcc-from-the-installation-path-is-it-allowed.
-        when Version.formula_optionally_versioned_regex(:gcc)
-          Regexp.union(%r{#{cellar_regex}/gcc}, %r{#{prefix_regex}/opt/gcc}) if OS.linux?
-        # binutils is relocatable for the same reason: https://github.com/Homebrew/brew/pull/11899#issuecomment-906804451.
-        when Version.formula_optionally_versioned_regex(:binutils)
-          %r{#{cellar_regex}/binutils} if OS.linux?
-        end
-        # rubocop:enable Homebrew/MoveToExtendOS
-
-        ignores.compact
+        cellar_regex = Regexp.escape(HOMEBREW_CELLAR)
+        go_regex = Version.formula_optionally_versioned_regex(:go, full: false)
+        Array(%r{#{cellar_regex}/#{go_regex}/[\d.]+/libexec})
       end
 
       sig { params(formula: Formula).void }
