+++ Library/Homebrew/dev-cmd/tests.rb
@@ -120,30 +120,13 @@ module Homebrew
           ]
           bundle_args << "--fail-fast" if args.fail_fast?
           bundle_args << "--profile" << args.profile if args.profile
-
-          # TODO: Refactor and move to extend/os
-          # rubocop:disable Homebrew/MoveToExtendOS
-          unless OS.mac?
-            bundle_args << "--tag" << "~needs_macos" << "--tag" << "~cask"
-            bundle_args << "--tag" << "~needs_homebrew_core" if ENV["CI"]
-            bundle_args << "--tag" << "~needs_svn" unless args.online?
-
-            files = files.grep_v(%r{^test/(os/mac|cask)(/.*|_spec\.rb)$})
-          end
-
-          unless OS.linux?
-            bundle_args << "--tag" << "~needs_linux" << "--tag" << "~needs_systemd"
-            files = files.grep_v(%r{^test/os/linux(/.*|_spec\.rb)$})
-          end
-          # rubocop:enable Homebrew/MoveToExtendOS
-
           bundle_args << "--tag" << "~needs_arm" unless Hardware::CPU.arm?
-
           bundle_args << "--tag" << "~needs_intel" unless Hardware::CPU.intel?
-
           bundle_args << "--tag" << "~needs_network" unless args.online?
-
           bundle_args << "--tag" << "~needs_ci" unless ENV["CI"]
+
+          bundle_args = os_bundle_args(bundle_args)
+          files = os_files(files)
 
           puts "Randomized with seed #{seed}"
 
@@ -170,6 +153,41 @@ module Homebrew
       end
 
       private
+
+      sig { params(bundle_args: T::Array[String]).returns(T::Array[String]) }
+      def os_bundle_args(bundle_args)
+        # for generic tests, remove macOS or Linux specific tests
+        non_linux_bundle_args(non_macos_bundle_args(bundle_args))
+      end
+
+      sig { params(bundle_args: T::Array[String]).returns(T::Array[String]) }
+      def non_macos_bundle_args(bundle_args)
+        bundle_args << "--tag" << "~needs_homebrew_core" if ENV["CI"]
+        bundle_args << "--tag" << "~needs_svn" unless args.online?
+
+        bundle_args << "--tag" << "~needs_macos" << "--tag" << "~cask"
+      end
+
+      sig { params(bundle_args: T::Array[String]).returns(T::Array[String]) }
+      def non_linux_bundle_args(bundle_args)
+        bundle_args << "--tag" << "~needs_linux" << "--tag" << "~needs_systemd"
+      end
+
+      sig { params(files: T::Array[String]).returns(T::Array[String]) }
+      def os_files(files)
+        # for generic tests, remove macOS or Linux specific files
+        non_linux_files(non_macos_files(files))
+      end
+
+      sig { params(files: T::Array[String]).returns(T::Array[String]) }
+      def non_macos_files(files)
+        files.grep_v(%r{^test/(os/mac|cask)(/.*|_spec\.rb)$})
+      end
+
+      sig { params(files: T::Array[String]).returns(T::Array[String]) }
+      def non_linux_files(files)
+        files.grep_v(%r{^test/os/linux(/.*|_spec\.rb)$})
+      end
 
       sig { returns(T::Array[String]) }
       def changed_test_files
@@ -250,3 +268,5 @@ module Homebrew
     end
   end
 end
+
+require "extend/os/dev-cmd/tests"
