+++ Library/Homebrew/extend/os/dev-cmd/tests.rb
@@ -1,0 +1,5 @@
+# typed: strict
+# frozen_string_literal: true
+
+require "extend/os/linux/dev-cmd/tests" if OS.linux?
+require "extend/os/mac/dev-cmd/tests" if OS.mac?
