+++ Library/Homebrew/extend/os/mac/diagnostic.rb
@@ -425,8 +425,8 @@ module OS
         end
 
         def check_deprecated_caskroom_taps
-          tapped_caskroom_taps = Tap.select { |t| t.user == "caskroom" || t.name == "phinze/cask" }
-                                    .map(&:name)
+          tapped_caskroom_taps = ::Tap.select { |t| t.user == "caskroom" || t.name == "phinze/cask" }
+                                      .map(&:name)
           return if tapped_caskroom_taps.empty?
 
           <<~EOS
