+++ Library/Homebrew/extend/os/tap.rb
@@ -1,0 +1,4 @@
+# typed: strict
+# frozen_string_literal: true
+
+require "extend/os/mac/tap" if OS.mac?
