+++ Library/Homebrew/resource.rb
@@ -4,7 +4,7 @@
 require "downloadable"
 require "mktemp"
 require "livecheck"
-require "extend/on_system"
+require "on_system"
 
 # Resource is the fundamental representation of an external resource. The
 # primary formula download, along with other declared resources, are instances
