+++ Library/Homebrew/tap.rb
@@ -1079,15 +1079,12 @@ class Tap
   # All locally installed and core taps. Core taps might not be installed locally when using the API.
   sig { returns(T::Array[Tap]) }
   def self.all
-    cache[:all] ||= begin
-      core_taps = [
-        CoreTap.instance,
-        # The conditional is valid here because we only want the cask tap on macOS.
-        (CoreCaskTap.instance if OS.mac?), # rubocop:disable Homebrew/MoveToExtendOS
-      ].compact
+    cache[:all] ||= installed | core_taps
+  end
 
-      installed | core_taps
-    end
+  sig { returns(T::Array[Tap]) }
+  def self.core_taps
+    [CoreTap.instance].freeze
   end
 
   # Enumerate all available {Tap}s.
@@ -1517,3 +1514,5 @@ class TapConfig
     Homebrew::Settings.delete key, repo: tap.path
   end
 end
+
+require "extend/os/tap"
